\name{doseData}
\alias{doseData}
\docType{data}
\title{Dose Data Example}
\description{
  Simulated dose data; a vector of length 20
  with 4 observations for each of 5 doses.
}
\usage{data(doseData)}
\format{
  The format is:
 num [1:20] 1 1 1 1 2 2 2 2 3 3 ...
}
\examples{
  data(doseData)
  doseData
}
\keyword{datasets}
