\name{timeData}
\alias{timeData}
\docType{data}
\title{Time Trend Data Example}
\description{
  Simulated time data; a vector of length 15
  with 3 observations for each of 5 time points.
}
\usage{data(timeData)}
\format{
  The format is:
 num [1:15] 1 1 1 1 2 2 2 2 3 3 ...
}
\examples{
  data(timeData)
  timeData
}
\keyword{datasets}
