\name{doseData}

\alias{doseData}

\docType{data}


\title{Dose Data Example}


\description{
Dose data; a vector of length 12
 with 3 observations for each of 4 doses.
}


\usage{data(doseData)}


\format{
  The format is:
 num [1:12] 1 1 1 2 2 2 3 3 ...
}


\examples{

data(doseData)
  
doseData

}


\keyword{datasets}
