\name{esData}
\alias{esData}
\docType{data}
\title{Example Data as ExpressionSet}
\description{
  Example Data under the form of an ExpressionSet.
  The actual data are the same as \code{geneData} and
  \code{doseData}.
}
\usage{data(esData)}
\seealso{\code{\link{geneData}}, \code{\link{doseData}}}
\examples{
  data(esData)
  esData
}
\keyword{datasets}
