\name{geneData}
\alias{geneData}
\docType{data}
\title{Gene Expression Data Example}
\description{
  Simulated gene expression data as a numerical matrix; 
  the expression values reflect 5 monotonically increasing
  and 5 monotonically decreasing profiles with 10 observations
  for each profile.
}
\usage{data(geneData)}
\format{
  The format is:
 num [1:100, 1:20] 0.598 2.939 0.543 1.526 0.912 ...
}
\examples{
  data(geneData)
  head(geneData)
}
\keyword{datasets}
